Use scikit-build to simplify proper Python package installation.
Using skbuild directly allows us to simplify our setup.py file and more
easily install egg-compliant packages.

It seems really hard to get both `pip` and bare setup.py to both work
with skbuild. We also need to figure out whether we want to allow or
encourage "develop" or "editable" installations. If so, we would
probably want them to point to the _skbuild staging area rather than
creating a `gmx` directory in the cwd of setup.py.

Since we want CMake installs to have proper egg-info or dist-info, we
might want to explicitly call the egg_info command from within CMake
and nowhere else. Note that supporting both full CMake installation
and setup.py installation is not quite the intended use case of skbuild,
and we should refer to the scikit-build test cases for CMake / binary
packages rather than "hybrid" or "pure".

In the end, we probably want to encourage direct use of CMake or
setup.py over `pip` because of the clearer access to cmake-related
options and flags and the somewhat arcane cache management of `pip`.

We also need to figure out the right way to tell skbuild that the binary
package has 'requirements', 'setup_requires', or whatever.

To do: the option to build GROMACS as part of the gmxapi installation
needs to be reenabled.