simulation input modification
For 0.0.7, a `modify_input` node
is inserted in the execution graph between the `load_tpr` and `md` nodes.
Later changes will insert additional work elements in the workflow, but
this can wait for schema 0.2. User interface is unchanged from 0.0.6.1.

The TPR file specified as the input source for the operation is copied
and updated by the framework in a temporary area as an implementation
detail. The original TPR file is left untouched.

API functionality tasks:

- [ ] Add TPR and checkpoint file objects to gmx.core
- [ ] Add "microstate" abstraction to gmx.core.
- [ ] Read nsteps from TPR
- [ ] Read current step from microstate
- [ ] set nsteps in TPR
- [ ] write new TPR file
- [ ] insert graph node to transform tpr file from source to tpr file with
  updated parameters for MD node