Begin GROMACS / gmxapi compatibility library.
Needs a bit more linking consideration. Probably easiest to make a
compiled library that links against gmxapi, but will have to use gromacs
headers and rely on transitively linked symbols from libgromacs (e.g.
for constructor and destructor of t_inputrec).

For GROMACS 2019 / gmxapi 0.0.7, rely on a static mapping of input
parameters to libgromacs data structures and members, but implement with
abstractions that allow for (first) extensibility and (second) more
elaborate infrastructure to express and handle the parameters.