Consolidate logic for whether to use ModularSimulator.
There appeared to be inconsistent logic in runner.cpp about whether
the ModularSimulator would be used, but the apparent inconsistency was
actually just the result of equivalent logic being performed in
multiple places, and unclear ownership of an environment variable. This
change should clarify some logic in runner.cpp and simplify the behavior
of the SimulatorBuilder.

* Reference GMX_DISABLE_MODULAR_SIMULATOR environment variable in a
  single location.
* Use a single utility function to determine whether to use
  ModularSimulator, and call in Mdrunner::mdrunner().
* Refactor SimulatorBuilder to perform an action based solely on its
  arguments, rather than to override the client-provided arguments based
  on an environment variable.

Change-Id: I775ef82461f30b79b4412a473e5a1aa66270e2d5